+++ src/pipecat/transports/local/audio.py
@@ -37,6 +37,14 @@ class LocalAudioInputTransport(BaseInputTransport):
     def read_raw_audio_frames(self, frame_count: int) -> bytes:
         return self._in_stream.read(frame_count, exception_on_overflow=False)
 
+    async def start(self):
+        await super().start()
+        self._in_stream.start_stream()
+
+    async def stop(self):
+        await super().stop()
+        self._in_stream.stop_stream()
+
     async def cleanup(self):
         # This is not very pretty (taken from PyAudio docs).
         while self._in_stream.is_active():
@@ -59,6 +67,14 @@ class LocalAudioOutputTransport(BaseOutputTransport):
 
     def write_raw_audio_frames(self, frames: bytes):
         self._out_stream.write(frames)
+
+    async def start(self):
+        await super().start()
+        self._out_stream.start_stream()
+
+    async def stop(self):
+        await super().stop()
+        self._out_stream.stop_stream()
 
     async def cleanup(self):
         # This is not very pretty (taken from PyAudio docs).